truely going zen mode
When entering goyo in tmux, previously, the tmux status bar persisted
but now with these two lines that are added the tmux bar will be
disabled while entering goyo and enabled when leaving goyo

Now that those two shell commands are being executed from vim it slows
down entering and leaving goyo mode, it is notice-able amount of time
but for now I'll manage.

There are plugins that will run shell commands asynchronously like
<https://github.com/skywind3000/asyncrun.vim> this one, I tried it, it
will make entering and leaving goyo normal as before i.e. instantaneous
while there will be a little bit of delay for the tmux status bar to
appear and disappear.

There are also other plugins written in lua for nvim which quiet doesn't
work as intended, though they might be better in design and  stuff but
the feel is just not right compared to goyo for vim, so sticking to vim.
- <https://github.com/Pocco81/true-zen.nvim>
- <https://github.com/folke/zen-mode.nvim>