adding configs for bash
with reference to 4e5833ee I was looking to set the aliases that I have
set in `.bash_aliases` in fish shell, which isn't POSIX compliant and
had a different syntax of its own, which I'm not interested in learning

Also for the nvm to work I had to add plugins to fish shell so that it
can understand nvm scripts which were written in bash.

I have a choice of going to zsh rather than bash but the last time I used
zsh I didn't had a good experience with it. Anyway I'll work for a while
in bash and figure out if it is nice enough to use or if I should use
zsh, because I know there will be no syntax highlighting and auto
complete in bash which is true for zsh.

Now that I have switched to bash, `z` the directory jumper has to be
added, I had one for fish which I don't think will work in bash