buf word count, brace highlight
The status line is modified to display word count in the current buffer,
it will update automatically as I type text in the buffer which is a nice
edition to have.

The matching pairs of brackets were being highlighted before which was
annoying to look at while I moved my cursor, sometimes I were unable to
find the cursor as to on which brace the cursor is, because both opening
and ending were highlighted in same color.

The braces are highlighted with the help of a plugin that comes by
default in vim which I have disabled in this commit, now that the plugin
is not being loaded there is no highlights of braces and it looks good
now.