nvim tabline just got better
I was not using tabline before and not even using tabs in nvim as well
as vim, now I want to use them so I have added this config in lualine
which adds a tabline with the same theme as the statusline.

In this tabline I have added tab numbers on the right side and a
filename thing on the left.

I wanted the tabline to show only when I have more than one tabs open
which is not the case with lualine, for which I added the plugin luatab
which shows the tabline when I have 2 or more tabs, but I'm not using it
as the theme and its looks are not good.