tweaking statusline some more
w.r.t previous commit (945bb40) I didn't notice this but there were two
places where filename was being displayed once in the status line and
once in the tabline. I have now remove the filename from the statusline

I have also added back the git diff and diagnostics, so that I don't
have to use `git diff` again and again to check how many changes I made.
And the diagnostics are to have a look at them at a glance.

In the status line at the rightmost position there was this thing which
showed which line and which column the cursor is on. I'm not finding it
necessary anymore as I can look at the line number directly and column
number, I never look at it.