indexing from 1, not 0
Big new addition is that the window numbers and pane numbers start from
1 and not 0, which makes it little more easy to select (keyboard layout)
0 is all the way at the end.

Also stopped renaming of windows automatically so that I can set the
window names as I want them to be.

Two more minor additions, unbinding the previous pane switching keys,
now they are free and can be used for other stuff.
Sending <C-a> key if required.