using leader instead of `;`
Since I have disabled h,j,k,l, I'm relying on `f` and `t` for movement
in a line and using `;` and `,` a lot, while `,` is ok but `;` is not
in nvim, as I have it mapped for telescope related things, as nvim will
wait for a while to check if some other key is being pressed, if not
then it will perform the functionality of `;` and this causes a delay, I
think 200ms which I had set in set.lua file

So in order to get rid of the delay all of the keymaps starting with `;`
had to be removed. And after changing those to use `leader` key instead
there was a collision with other keys and hence had to rework some of
those as well.

All keys are followed by <Leader>
- f : [F]iles
- g : [G]rep
- b : [B]uffers
- t : [T]ags
- r : [R]ename
- v : Explorer, just because `v` is conveniently below `f` that is used
      for file
- d : [D]iagnostics
- c : [C]ode actions
- a : [A]dd file (harpoon)
- q : setloclist
      - Might change to `l` for [L]ocation list
- e : open_float (harpoon)
      - Might change to `Q` as it is free and `q` is above `a` for both
        harpoon keys to be at one place
- ; : resume (telescope)

Now that `f` is taken for [F]iles, the key for format has to go and as
it is a less used thing, <F3> is here to help.