fix: using defaults
tmux is my default shell and so it handles all the scroll buffers which
means the scroll buffer of alacritty is never used so disabling it.

setting env is also handled within tmux, so removed it as well.

cursor shape and other stuff has sane defaults, so removed it.

using bigger font size and lesser padding.