fix: wget certificate issue
`wget' is installed via brew as it doesn't come with MacOS, it doesn't
use the MacOS keychain and looks for certificate files in some location
that MacOS doesn't have (might have removed in MacOS v13).

By using `--ca-certificate' option we can explicitly specify to use
these certificate files, also `ca-certificates' has to be downloaded.

This `wget' alias requires `brew' command, to know its prefix (not to
hardcode the path of `ca-certificates') and hence the eval line of brew
should be executed before we can set this alias (to be able to use
`brew --prefix'). And for this reason, loading aliases at the end.

And yeah using a TUI client for matrix, yes I joined there.