feat(nvim): setting `tectonic' as LaTeX compiler
Firstly now there is no need for `:Tectonic' usercmd as this commit adds
the ability to use `:make' with `.tex' files.

Determining if I need the version 1 or 2 interface of the compiler i.e.
`tectonic' I can set appropriate flag and set different options for
`makeprg'

`tectonic' outputs the file names ignore the `src/' directory which is
annoying and the quickfix list opens new buffer if the specified file
doesn't exist, which was always the case because all files existed in
`src/'. To fix this there is no way of adding the leading `src/' to the
fine name before they are added to quickfix list. There are methods such
as `setqflist' but that would be more manual work.

There is an option of using "filters" for filtering the output before it
is read and processed according to the `errorformat'. This commit also
adds a filter which takes in the output of `tectonic' and adds the
leading `src/' to the file names and outputs it.

When `:make' is called the `tectonic' binary is called piping its output
to this filter and the output of this filter is read according to the
specified `errorformat' and the quickfix list is populated.

The `tectonic' compiler file is written in vimscript as I do not yet
know how to write all of it in lua, ideally this should be rewritten in
lua.

The `errorformat' is defined as:
  %t        - e | w
  %*[^:]:   - anything upto :
  %f        - filename
  %l        - line number
  %m        - message
(spaces have to be escaped using '\ ')

Or we can have this format that will only care about the error and
warning messages and ignores everything else.
```vimscript
CompilerSet errorformat=
                \%t%*[^:]:\ %f:%l:\ %m,
                \%-G%.%#
```
  %-G       - Ignore this
  %.%#      - anything upto end of line
  Leading `\' is to escape the newline.... vimscript thing
  errorformat lines are separated by `,'