feat(git): paginate output properly
Reading `core.pager' section in `git-config(1)' says that `git' will set
the `LESS' environment variable to `FRX' and call `less'.

Reading what these options mean in `less(1)' reveals that if `-X' is
present then upon exit, less does not clear the screen. And further
reading of how command line arguments and environment variables are read
reveals that adding `+' in options negates the effect if the same option
is present in `LESS' environment variable.

The `core.pager' also reveals that if `LESS' is already defined then it
will not be modified and so another way of achieving this would be to
export `LESS=R' instead of what is added in this commit.