fix(zsh): save starship prompt loading time
99% of the time the pwd is enough and only 1% of the time I look at the
other stuff that `starship` provides, that was the reason for `^P`
keybinding to exist in the first place.

While 99% of the times I was not in need of the starship prompt and was
still loading it which takes around 12ms. By using the builtin
mechanisms of prompt initially that time can be saved.

The mechanism to check if `starship` is loaded is to check if the
environment variable is set:
`https://stackoverflow.com/questions/42655304`