remove vim config from here
Instead use a separate repo just for vim `github.com/SyedFasiuddin/.vim`
Why? Because `vim` will be present everywhere and to have my personal
config for vim on that machine, I will need to get my entire `.dotfiles`
and then pick my `.vim` folder, so instead it is a good idea to have
just my `.vim` in separate repo and just get it.