fix: brew placement in path
Why?
Because MacOS comes with python installed current version is 3.9.6 and
brew installes python as a dependency for other packages that I had
installed. Brew can remove this python anytime or upgrade which will all
sorts of problems if I use it, by moving the brew/bin path later than
the system paths I can use the system provided python just in case I
need it, and project specific versions can be installed with nix.

Python is the only reason to do this, if something breaks then I might
change this later.