fix(zsh/nix): manage zsh config with nix
My config is based on `stow`, it used to manage all the links for me but
now `home-manager` does.

The `home.file` is used to link the files similar to stow but this is
more explicit as to saying which file is linked where. By managing links
this way I can get rid of `stow` completely.

Why rename `.zshrc`?
Because home-manager creates that file and fills it with all the config
options that are explicitly stated here or are added by some other
module such as `fzf` will add its loading configuration automatically.

Since `dotDir` option sets the `$ZDOTDIR` variable we could get rid of
`.zshenv`