fix: remove project specific binaries
`mdbook` is only needed while writing any books, so it can be installed
via a `flake.nix` or `shell.nix` in that project. Same is true for
`presenterm` and `universal-ctags`.

`texinfo` package doesn't install the `info` command but instead it only
installed those commands that are used to build the `.info` files. The
`info` command however is provided by this `texinfoInteractive` package.

`wiki-tui` is the least used thing, wikipedia in browser is better, and
this tui version doesn't display a lot of things.

Adding `irssi` to dive into IRC chats, its configuration will be added
soon.