feat(nix): manage all config files and scripts
All these configs were single files, so decided to have a `config/`
folder to contain all single file configs and when they have more files
like in the case of `tmux/` and `nvim/` separate them.

I was not sure about the name `config/` but decided to go with it
anyway.